requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
httpx==0.25.2
celery==5.3.4
python-telegram-bot==20.7
aiofiles==23.2.1
Pillow==10.1.0
pandas==2.1.4
openpyxl==3.1.2
jinja2==3.1.2
python-dateutil==2.8.2
pytz==2023.3
